feat(release): v3.0.2 1. PanasonicMewtocolOverTcp协议修复在站号输入大于128时，解析的异常，并且修复在报文中输出的站号为小写字母e的bug。 2. MelsecA1ENet: 修复在读取bool数组时，解析不正确的bug。 3. SiemensS7Net: 新增读写WString字符串类型的功能，支持中文的读写操作。 4. 本软件已经申请软件著作权，软著登字第5219522号，任何盗用软件，破解软件，未经正式合同授权而商业使用均视为侵权。

+++ src/HslCommunicationDemo/FormPanasonicMewtocolOverTcp.java
@@ -116,7 +116,7 @@ public class FormPanasonicMewtocolOverTcp extends JDialog {
                 try {
                     plc.setIpAddress(textField1.getText());
                     plc.setPort(Integer.parseInt(textField2.getText()));
-                    plc.setStation(Byte.parseByte(textField4.getText()));
+                    plc.setStation(Integer.parseInt(textField4.getText()));
 
                     OperateResult connect = plc.ConnectServer();
                     if(connect.IsSuccess){
